+++ src/commands/produce/channel.utils.ts
@@ -1,0 +1,15 @@
+import { Channel } from "amqplib";
+
+/**
+ * Helper function which can be used to block execution (sending further messages)
+ *   till the 'drain' event is emit form the channel.
+ *
+ * @param channel The channel to wait for
+ */
+export async function waitForDrain(channel: Channel): Promise<void> {
+  return new Promise((resolve) => {
+    channel.once("drain", () => {
+      resolve();
+    });
+  });
+}