+++ src/commands/produce/index.test.ts
@@ -1,0 +1,27 @@
+import { createCommand, Command } from "commander";
+import { buildProduceCommand } from "./";
+import { CommandFactoryFn } from "../common";
+
+describe("Produce Command", () => {
+  let consume: Command;
+
+  beforeEach(() => {
+    consume = buildProduceCommand(createCommand as CommandFactoryFn);
+  });
+
+  test("it returns a command object as a result", () => {
+    expect(consume).toBeInstanceOf(Command);
+  });
+
+  test("it defines the to-queue sub-command", () => {
+    expect(consume.commands.some((cmd) => cmd.name() === "to-queue")).toEqual(
+      true
+    );
+  });
+
+  test("it defines the to-exchange sub-command", () => {
+    expect(
+      consume.commands.some((cmd) => cmd.name() === "to-exchange")
+    ).toEqual(true);
+  });
+});