+++ src/commands/produce/types.ts
@@ -1,0 +1,7 @@
+import { OnMessageFn } from "../../utils/io";
+import { Debugger } from "debug";
+
+export type InputProvider = (
+  onMessage: OnMessageFn,
+  logger: Debugger
+) => Promise<void>;